+++ CHANGELOG.md
@@ -9,6 +9,13 @@ genauso wie an [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 Aus Gründen der Übersichtlichkeit sind bei älteren Versionen die einzelnen Patch-Versionen nicht extra aufgeführt, sondern in der Minor-Version.
 
 
+## [Unreleased]
+
+### Changed
+
+- Teildatensatz.getFeld(int) auf maximale Feld-Anzahl beschränkt
+
+
 ## [7.0.1] - 2024-01-26
 
 ### Fixed